Move the ArtifactManifest out of artifacts.py
